backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.7.0
google-cloud-run>=0.10.0
google-cloud-build>=3.22.0
google-auth>=2.29.0
